Add files via upload
Pie-in-the-Sky, is a mobile app that is used for bidding for IPL matches legally. The app shows the match details which includes the playing team, the venue of the match and current standing of the teams on the points table. It will display the winner at the end of the match and update the team standings in the tournament and bidder points table.